requirements.txt
flask==2.0.1
gunicorn[gevent]
requests
Flask-HTTPAuth
passlib[bcrypt]
pyjwt>=2
pyyaml
postgres
password-strength
brotli
apscheduler
html5lib
jsonschema==3.2.0
python-dateutil
beautifulsoup4
geopandas
geoplot
Shapely
pygeos==0.10.2
python-Levenshtein==0.12.2
wheel
pytest
pytest-xdist
pytest-timeout
numpy==1.21.2
